[COMM-628] Remove arrows from vote buttons
Rather than adding

* Unicode Character 'UPWARDS BLACK ARROW' (U+2B06)
* Unicode Character 'DOWNWARDS BLACK ARROW' (U+2B07)

to the vote buttons using `::before` and a font to go with those
symbols we want to use SVG to display those arrows.  To style those
icons I'm also removing the `::before` selectors for the remaining
rules that deal with styling (hover highlight etc) for these elements.

The modification of the `font-size`, `display` and `flex-direction`
was to fix a centering/alignment issue with this new approach.

I've copied over this file:

https://github.com/zendeskgarden/svg-icons/blob/master/src/16/chevron-down-fill.svg

From zendeskgarden.  For the upwards pointing caret I've added an SVG
transformation to the root `<svg>` element to mirror it along the
horizontal axis.  Otherwise the two elements are the same.

I've also removed the `outline: none;` rule.