fix: apply end user conditions for checkbox fields correctly
The end user conditions for checkbox fields with unselected values were not being applied
correctly.
This was due to the fact that the conditions were not being applied at the first page
load, when the field value has a default value of null,
because we were comparing it with the condition value of false.