fix: fixed upload of some type of attachments
We were determining the mime type for attachments in the theme using
the `mime` library, and passing this info to the Upload Attachments API
in the `Content-Type` header.

This was implemented in #520 to support some specific files, as
suggested by the team that owns the endpoint, but it still causes some
issues with some files (namely `.log` files not encoded as UTF-8).

The team advises to just pass `application/octet-stream` as type if the
browser doesn't detect the type automatically. This solves the
upload issue for all the problematic files we know so far, and
we can stop bundling a library just for determining the mime type of the
uploaded files.