feat: improve form controls border contrast for accessibility
The default border colors for form controls in the default theme in Zendesk Garden
don't meet the minimum 3:1 contrast ratio against the background.
We apply a custom theme override to increase the contrast of these borders,
following the guidelines provided by the Garden team in an internal doc.
The guidelines suggest adjusting the border colors by applying offsets,
the approach has been extended to take into accounts all form controls.

By default, Garden sets the `grey.300` color as the default border color
in the light theme, and then it is increased by 100 for form controls,
which is not sufficient, providing only 2:1 contrast ratio.
To reach the 3:1 ratio, we need to use `grey.600` and so
we apply an offset of 300.

We also enhance the hover state border color, which by default is less
prominent when the border color is darker.

We also update the `$high-contrast-border-color` Sass variable
to match the new border color used in the theme override.